Refactored crash detection a bit
Fixes test-run's hang when server running 'box/box.lua' killed by a
signal (at end of the test).

Fixes cases like double/triple crash reporting (non-default server
crash, then default one, then the results mismatch). Fixes cases where
a log file was didn't shown.

Added basic support for detection crash of non-default servers in Python
tests (related to #35), but some clean up like in LuaTest.execute()
after `execfile()` call would be better to add. Also I don't sure
raising Python exception from the test will processed appropriatelly.

Reimplemented startup fail reporting (fixes #44), more cases of failing
tests (default/non-default servers; app/tarantool/python tests) should
be processed correctly.

Non-default server's crashes detected now by test-run, so test-run
provided ability for Python tests to allow server silently failing. For
now the only test expects from server to fail is
'replication-py/cluster.test.py'. It needs to be updated with:

```
 failed.vardir = server.vardir
 failed.rpl_master = replica
 failed.name = "failed"
+failed.crash_expected = True
 try:
     failed.deploy()
 except Exception as e:
```

+++ lib/test.py
@@ -17,6 +17,13 @@ from lib.utils import check_valgrind_log, print_tail_n
 color_stdout = Colorer()
 
 
+class TestExecutionError(OSError):
+    """ When we know we fail, but already reported the fail in an another
+    greenlet.
+    """
+    pass
+
+
 class TestRunGreenlet(gevent.Greenlet):
     def __init__(self, green_callable, *args, **kwargs):
         self.callable = green_callable
@@ -105,12 +112,20 @@ class Test:
         self.run_params = params
         self.conf_name = conf_name
 
+        # filled in execute() when a greenlet runs
+        self.current_test_greenlet = None
+
+        # prevent double/triple reporting
+        self.is_crash_reported = False
+
     def passed(self):
         """Return true if this test was run successfully."""
         return self.is_executed and self.is_executed_ok and self.is_equal_result
 
     def execute(self, server):
-        pass
+        # Note: don't forget to set 'server.current_test = self' in
+        # inherited classes. Crash reporting relying on that.
+        server.current_test = self
 
     def run(self, server):
         """Execute the test assuming it's a python program.
@@ -126,6 +141,9 @@ class Test:
             if os.path.exists(self.skip_cond):
                 sys.stdout = FilteredStream(self.tmp_result)
                 stdout_fileno = sys.stdout.stream.fileno()
+                # XXX: Not sure crash reporting is adequate on these tests,
+                # maybe we need to pass 'test_run_current_test' to their
+                # non-default servers.
                 execfile(self.skip_cond, dict(locals(), **server.__dict__))
                 sys.stdout.close()
                 sys.stdout = save_stdout
@@ -135,6 +153,8 @@ class Test:
                 self.execute(server)
                 sys.stdout.flush()
             self.is_executed_ok = True
+        except TestExecutionError:
+            self.is_executed_ok = False
         except Exception as e:
             traceback.print_exc(e)
             diagnostics = str(e)
@@ -171,15 +191,15 @@ class Test:
             color_stdout("[ fail ]\n", schema='test_fail')
 
             where = ""
-            if not self.is_executed_ok:
+            if not self.is_crash_reported and not self.is_executed_ok:
                 self.print_diagnostics(self.reject, "Test failed! Last 10 lines of the result file:\n")
                 server.print_log(15)
                 where = ": test execution aborted, reason '{0}'".format(diagnostics)
-            elif not self.is_equal_result:
+            elif not self.is_crash_reported and not self.is_equal_result:
                 self.print_unidiff()
                 server.print_log(15)
                 where = ": wrong test output"
-            elif not self.is_valgrind_clean:
+            elif not self.is_crash_reported and not self.is_valgrind_clean:
                 os.remove(self.reject)
                 self.print_diagnostics(server.valgrind_log, "Test failed! Last 10 lines of valgrind.log:\n")
                 where = ": there were warnings in valgrind.log"
